+++ Extensions/Announcers/DisqusAnnouncer/Properties/AssemblyInfo.cs
