+++ Extensions/Announcers/SampleAnnouncement/Properties/AssemblyInfo.cs
