+++ Extensions/Announcers/TwitterAnnouncer/Properties/AssemblyInfo.cs
