+++ Extensions/Announcers/UserVoiceAnnouncer/Properties/AssemblyInfo.cs
